Extract LSP commands
